Added parallel project handler

+++ Assets/scripts/RevenueManager.cs
@@ -2,15 +2,13 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class RevenueManager : MonoBehaviour
+public class RevenueManager: MonoBehaviour
 
 {
-    public void increaseRevenue(double increment)
-    {
+    public void increaseRevenue(double increment) {
         Controller.instance.revenue += increment;
 
-        if(Controller.instance.revenue >= Controller.instance.revenueMilestones[Controller.instance.GameLevel-1].Milestone)
-        {
+        if (Controller.instance.revenue >= Controller.instance.revenueMilestones[(int) Controller.instance.GameLevel - 1].Milestone) {
             Controller.instance.levelUp();
         }
     }
